add inner copy
