+++ jupyterlab-dataops/requirements/requirements-bigquery.txt
@@ -1,0 +1,2 @@
+dbt-bigquery==1.0.0
+google-cloud-bigquery[bqstorage,pandas]==2.32.0
