+++ jupyterlab-dataops/requirements/requirements.txt
@@ -1,0 +1,6 @@
+dbt-core==1.0.4
+copier==5.1.0
+acryl-datahub[datahub-rest]==0.8.17.7
+data-pipelines-cli[docker,datahub,gcs,s3]==0.20.0
+jupyterlab-git==0.37.1
+jupyter-server-proxy==3.2.1
