+++ jupyterlab-dataops/resources/cloudbeaver/conf/.cloudbeaver.auto.conf
@@ -1,0 +1,6 @@
+{
+    CB_ADMIN_NAME: 'admin',
+    CB_ADMIN_PASSWORD: 'admin',
+    CB_SERVER_NAME: 'CloudBeaver DataOps',
+    CB_SERVER_URL: 'http://localhost:8978'
+}
