+++ jupyterlab-dataops/resources/cloudbeaver/conf/product.conf
@@ -1,0 +1,34 @@
+// Product configuration. Customized web application behavior
+// It is in JSONC format
+{
+    // Global properties
+    core: {
+        // User defaults
+        user: {
+            defaultTheme: "light",
+            defaultLanguage: "en"
+        },
+        app: {
+            // Log viewer config
+            logViewer: {
+                refreshTimeout: 3000,
+                logBatchSize: 1000,
+                maxLogRecords: 2000,
+                maxFailedRequests: 3
+            }
+        },
+        authentication: {
+            primaryAuthProvider: 'local'
+        }
+    },
+    // Notifications config
+    core_events: {
+        notificationsPool: 5
+    },
+    plugin_data_spreadsheet_new: {
+        hidden: false
+    },
+    plugin_data_export: {
+        disabled: false
+    }
+}
