+++ jupyterlab-dataops/resources/vsc/settings.json
@@ -1,0 +1,6 @@
+{
+  "files.associations": {
+    "*.sql": "jinja-sql"
+  },
+  "python.pythonPath": "/opt/conda/bin/python"
+}
